Removed partial population of the "measurement_information" dictionary in PulseBlockEnsembles and PulseSequences altogether when creating these objects via graphical pulse editors. In that sense also got rid of the "use_as_tick" parameter in PulseBlocks.
