+++ gui/pulsed/pulsed_maingui.py
@@ -20,7 +20,6 @@ Copyright (c) the Qudi Developers. See the COPYRIGHT.txt file at the
 top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi/>
 """
 
-import inspect
 import numpy as np
 import os
 import pyqtgraph as pg
