+++ gui/pulsed/ui_pulse_editor.ui
@@ -1069,12 +1069,12 @@ last</string>
   <customwidget>
    <class>BlockEditor</class>
    <extends>QTableView</extends>
-   <header>gui.pulsed.pulse_editors_new.h</header>
+   <header>gui.pulsed.pulse_editors.h</header>
   </customwidget>
   <customwidget>
    <class>EnsembleEditor</class>
    <extends>QTableView</extends>
-   <header>gui.pulsed.pulse_editors_new.h</header>
+   <header>gui.pulsed.pulse_editors.h</header>
   </customwidget>
  </customwidgets>
  <tabstops>
